GradleRecipesForAndroid/ch05/Demo2/app/src/main/java/com/example/luoheng/demo/MainActivity.java
package com.example.luoheng.demo;

import android.content.Intent;
import android.support.v7.app.AppCompatActivity;
import android.os.Bundle;
import android.util.Log;
import android.view.View;
import android.widget.EditText;

public class MainActivity extends AppCompatActivity {

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_main);
        Log.d("luoheng", "Demo------" + Utils.printPTID() + ",pkg " + getPackageName());

        final EditText editText = findViewById(R.id.et_text);
        findViewById(R.id.btn).setOnClickListener(new View.OnClickListener() {
            @Override
            public void onClick(View v) {
                String name = editText.getText().toString();
                Intent intent = new Intent(MainActivity.this, WelcomeActivity.class);
                intent.putExtra("name", name);
                startActivity(intent);
            }
        });
    }
}
